Base64 buffer-first implementation
`ByteString` relies on Buffer implementation for Base64 capabilities.

`BufferedSource` and `BufferedSink` API for Base64.
