+++ okio/src/jvmMain/kotlin/okio/Buffer.kt
@@ -330,6 +330,12 @@ actual class Buffer : BufferedSource, BufferedSink, Cloneable, ByteChannel {
   @Throws(EOFException::class)
   override fun readUtf8CodePoint(): Int = commonReadUtf8CodePoint()
 
+  @Throws(EOFException::class)
+  override fun readBase64(): String = commonReadBase64()
+
+  @Throws(EOFException::class)
+  override fun readBase64Url(): String = commonReadBase64Url()
+
   override fun readByteArray() = commonReadByteArray()
 
   @Throws(EOFException::class)
@@ -447,6 +453,9 @@ actual class Buffer : BufferedSource, BufferedSink, Cloneable, ByteChannel {
 
   actual override fun writeHexadecimalUnsignedLong(v: Long): Buffer =
     commonWriteHexadecimalUnsignedLong(v)
+
+  actual override fun writeBase64(string: String): Buffer =
+    commonWriteBase64(string)
 
   internal actual fun writableSegment(minimumCapacity: Int): Segment =
     commonWritableSegment(minimumCapacity)
