+++ okio/src/nativeMain/kotlin/okio/Buffer.kt
@@ -132,6 +132,10 @@ actual class Buffer : BufferedSource, BufferedSink {
 
   override fun readUtf8CodePoint(): Int = commonReadUtf8CodePoint()
 
+  override fun readBase64(): String = commonReadBase64()
+
+  override fun readBase64Url(): String = commonReadBase64Url()
+
   override fun readByteArray(): ByteArray = commonReadByteArray()
 
   override fun readByteArray(byteCount: Long): ByteArray = commonReadByteArray(byteCount)
@@ -191,6 +195,9 @@ actual class Buffer : BufferedSource, BufferedSink {
 
   actual override fun writeHexadecimalUnsignedLong(v: Long): Buffer =
     commonWriteHexadecimalUnsignedLong(v)
+
+  actual override fun writeBase64(string: String): Buffer =
+    commonWriteBase64(string)
 
   override fun write(source: Buffer, byteCount: Long): Unit = commonWrite(source, byteCount)
